Junção de módulos em um único
- Remoção de modulo exclusivo para registro e login
- Criação de modulo usuário
