+++ setup/settings.py
@@ -21,8 +21,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "rest_framework",
-    "register",
-    "login"
+    "user"
 ]
 
 MIDDLEWARE = [
