+++ setup/urls.py
@@ -3,6 +3,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("", include('register.urls')),
-    path("", include('login.urls'))
+    path("", include('user.urls'))
 ]
