+++ user/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
-class RegisterConfig(AppConfig):
+class LoginConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "register"
+    name = "user"
